requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==2.4.1
attrs==25.3.0
base==0.0.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.2.1
comm==0.2.1
contourpy==1.2.1
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.0.1
fastapi==0.123.5
filelock==3.15.4
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.8
imageio==2.35.1
imgaug==0.4.0
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.4
limits==5.6.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
narwhals==2.4.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
pyotp==2.9.0
pyparsing==3.1.4
python-dateutil==2.8.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.7.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==73.0.1
shapely==2.0.6
six==1.16.0
slowapi==0.1.9
smmap==5.0.1
sounddevice==0.5.2
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
streamlit==1.49.1
sympy==1.13.2
tenacity==9.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.28
toml==0.10.2
torch==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.4
wrapt==2.0.1